Pull TGF graph serialization into I/O module
