+++ generative/io/mod.rs
@@ -1,7 +1,9 @@
 mod stdio;
+mod tgf;
 mod wkt;
 
 pub use stdio::{get_input_reader, get_output_writer};
+pub use tgf::{write_graph, GraphFormat};
 
 pub use self::wkt::{
     read_geometries, read_wkt_geometries, write_geometries, write_wkt_geometries, GeometryFormat,
