+++ generative/lib.rs
@@ -1,4 +1,5 @@
 pub mod dla;
 pub mod flatten;
+pub mod graph;
 pub mod io;
 pub mod triangulation;
