+++ generative/triangulation.rs
@@ -1,15 +1,7 @@
-use clap::ValueEnum;
 use geo::Point;
-use petgraph::visit::EdgeRef;
 use petgraph::{Directed, Undirected};
 
-use crate::io::write_wkt_geometries;
-
-type NodeData = Point;
-type EdgeWeight = ();
-type NodeIndex = usize;
-pub type DiGraph = petgraph::Graph<NodeData, EdgeWeight, Directed, NodeIndex>;
-pub type Graph = petgraph::Graph<NodeData, EdgeWeight, Undirected, NodeIndex>;
+use crate::graph::{GeometryGraph, NodeIndex};
 
 /// Calculate the Delaunay triangulation of the given point cloud
 pub fn triangulate(points: impl Iterator<Item = geo::Point>) -> Triangulation {
@@ -80,10 +72,10 @@ impl Triangulation {
         geo::Polygon::new(hull, vec![])
     }
 
-    pub fn digraph(&self) -> DiGraph {
+    pub fn digraph(&self) -> GeometryGraph<Directed> {
         let nodes = self.points.len();
         let edges = self.triangulation.halfedges.len();
-        let mut graph = DiGraph::with_capacity(nodes, edges);
+        let mut graph = GeometryGraph::with_capacity(nodes, edges);
 
         // Add all the nodes
         for (_i, point) in self.points.iter().enumerate() {
@@ -132,11 +124,12 @@ impl Triangulation {
         graph
     }
 
-    pub fn graph(&self) -> Graph {
+    pub fn graph(&self) -> GeometryGraph<Undirected> {
         let digraph = self.digraph();
         let nodes = self.points.len();
         let directed_edges = self.triangulation.halfedges.len();
-        let mut graph = Graph::with_capacity(nodes, directed_edges - self.triangulation.hull.len());
+        let mut graph =
+            GeometryGraph::with_capacity(nodes, directed_edges - self.triangulation.hull.len());
 
         // Add the nodes
         for (_i, node) in digraph.raw_nodes().iter().enumerate() {
@@ -176,7 +169,7 @@ impl Triangulation {
         longest
     }
 
-    pub fn urquhart(&self) -> Graph {
+    pub fn urquhart(&self) -> GeometryGraph<Undirected> {
         let mut graph = self.graph();
 
         // According to https://en.wikipedia.org/wiki/Urquhart_graph you can construct the Urquhart
@@ -196,68 +189,6 @@ impl Triangulation {
 
         graph
     }
-}
-
-#[derive(Debug, Clone, ValueEnum)]
-pub enum GraphFormat {
-    Tgf,
-    Wkt,
-}
-
-pub fn write_graph<Direction, W>(
-    writer: W,
-    graph: petgraph::Graph<NodeData, EdgeWeight, Direction, NodeIndex>,
-    format: &GraphFormat,
-) where
-    W: std::io::Write,
-    Direction: petgraph::EdgeType,
-{
-    match format {
-        GraphFormat::Tgf => write_graph_tgf(writer, graph),
-        GraphFormat::Wkt => write_graph_wkt(writer, graph),
-    }
-}
-
-fn write_graph_tgf<Direction, W>(
-    mut writer: W,
-    graph: petgraph::Graph<NodeData, EdgeWeight, Direction, NodeIndex>,
-) where
-    W: std::io::Write,
-    Direction: petgraph::EdgeType,
-{
-    // let (nodes, edges) = graph.into_nodes_edges();
-    for idx in graph.node_indices() {
-        let coord = graph
-            .node_weight(idx)
-            .expect("Got index to nonexistent node.");
-        let index = idx.index();
-        writeln!(writer, "{}\tPOINT({} {})", index, coord.x(), coord.y())
-            .expect("Failed to write node label");
-    }
-    writeln!(writer, "#").expect("Failed to write node/edge separator");
-    for edge in graph.edge_references() {
-        writeln!(
-            writer,
-            "{}\t {}",
-            edge.source().index(),
-            edge.target().index()
-        )
-        .expect("Failed to write edge");
-    }
-}
-
-fn write_graph_wkt<Direction, W>(
-    writer: W,
-    graph: petgraph::Graph<NodeData, EdgeWeight, Direction, NodeIndex>,
-) where
-    W: std::io::Write,
-    Direction: petgraph::EdgeType,
-{
-    let edges = graph
-        .edge_references()
-        .map(|e| geo::Line::new(graph[e.source()], graph[e.target()]))
-        .map(geo::Geometry::Line);
-    write_wkt_geometries(writer, edges);
 }
 
 #[cfg(test)]
